requirements.txt
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.5.6
botocore==1.8.20
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.8
django-allauth==0.34.0
django-crispy-forms==1.7.0
django-htmlmin==0.10.0
django-storages==1.6.5
django-wordpress-api==0.1.24
docutils==0.14
gunicorn==19.7.1
html5lib==1.0.1
iso8601==0.1.11
jmespath==0.9.3
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.11.1
requests-oauthlib==0.8.0
s3transfer==0.1.12
six==1.10.0
webencodings==0.5.1
